[refactor] Move legacy processors to processors/legacy/ and remove MultimodalSequence2SequenceProcessor
Legacy task-specific processor wrappers (Pose2Text, Video2Text, Features2Text,
Image2Text, Text2Text, Signwriting) are now under processors/legacy/ to make
clear they are backward-compatibility shims around the slot-based design.
A legacy/__init__.py re-exports all six classes, keeping the public API in
multimodalhugs.processors unchanged.

MultimodalSequence2SequenceProcessor is deleted — it is no longer used by any
processor after the step 7 refactor.

Update all test and asset imports to reference processors.legacy.* explicitly.
Update CLAUDE.md to reflect the new processor architecture.
