+++ multimodalhugs/processors/legacy/__init__.py
@@ -1,0 +1,6 @@
+from .pose2text_preprocessor import Pose2TextTranslationProcessor
+from .features2text_preprocessor import Features2TextTranslationProcessor
+from .video2text_preprocessor import Video2TextTranslationProcessor
+from .image2text_preprocessor import Image2TextTranslationProcessor
+from .text2text_preprocessor import Text2TextTranslationProcessor
+from .signwriting_preprocessor import SignwritingProcessor
