+++ multimodalhugs/processors/legacy/image2text_preprocessor.py
