+++ multimodalhugs/processors/legacy/signwriting_preprocessor.py
