+++ tests/test_data/test_processor_pose2text.py
@@ -3,7 +3,7 @@
 import torch
 from transformers.feature_extraction_utils import BatchFeature
 
-from multimodalhugs.processors.pose2text_preprocessor import (
+from multimodalhugs.processors.legacy.pose2text_preprocessor import (
     Pose2TextTranslationProcessor,
 )
